refactor: Move resend command to a package
